moved configuration
