+++ api/configuration/database/exceptions.py
