+++ api/configuration/email/configuration.py
