+++ api/configuration/email/variables.py
