+++ api/configuration/exceptions.py
