+++ api/services/email/email.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from typing import *
 from abc import ABC
-from configuration import Configuration
+from api.configuration import Configuration
 from abc import abstractmethod
 from shared.dependency.exceptions import DependencyNotFoundError
