+++ cli/cli.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
-from configuration import Configuration
-from configuration.environment.types import TEnvironment
-from configuration.variables import Variables
+from api.configuration import Configuration
+from api.configuration.environment.types import TEnvironment
+from api.configuration.variables import Variables
 from .options import configuration_options
 from . import ui
 from . import database
