+++ cli/database.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
-from configuration.environment.types import TEnvironment
+from api.configuration.environment.types import TEnvironment
+from api.configuration import Configuration
 from api.database import Database
-from configuration import Configuration
 from .options import configuration_options
 import asyncio
 from alembic import command
